+++ core/lib/cc/modcustom.go
@@ -85,14 +85,16 @@ func moduleCustom() {
 		}
 
 		// compress module files
-		tarball := WWWRoot + CurrentMod + ".tar.bz2"
-		CliPrintInfo("Compressing %s with bz2...", CurrentMod)
+		tarball := WWWRoot + CurrentMod + ".tar.xz"
+		CliPrintInfo("Compressing %s with xz...", CurrentMod)
 		path := fmt.Sprintf("%s/%s", config.Path, CurrentMod)
-		err = util.TarBz2(path, tarball)
+		err = util.TarXZ(path, tarball)
 		if err != nil {
 			CliPrintError("Compressing %s: %v", CurrentMod, err)
 			return
 		}
+		CliPrintInfo("Created %.4fMB archive (%s) for module '%s'",
+			float64(util.FileSize(tarball))/1024/1024, tarball, CurrentMod)
 
 		// tell agent to download and execute this module
 		checksum := tun.SHA256SumFile(tarball)